requirements.txt
# Core dependencies
python-dotenv>=1.0.0
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Vector store and embeddings
chromadb>=0.4.22
sentence-transformers>=2.3.1

# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
pydantic>=2.5.0
numpy>=1.26.0
pandas>=2.1.4

# Optional: for better logging
colorama>=0.4.6

